Add base command & update README
