+++ lib/lib/options.rb
