+++ lib/lib/post.rb
