+++ lib/lib/stats.rb
