+++ lib/tumblr_utilities.rb
@@ -21,5 +21,9 @@ require_relative('lib/tumblr_client.rb')
 require_relative('lib/options.rb')
 
 # load in our commands!
+# load this one first
+require_relative('command/base/iterate_through_posts.rb')
+
+# and then these
 require_relative('command/privatize_posts.rb')
 require_relative('command/update_community_labels.rb')
