+++ script.rb
